feat: add task succeed as separate function;
