--- .jest/setEnvVars.js
@@ -1,2 +1,0 @@
-process.env.GITHUB_API_URL = 'https://api.github.com';
-process.env.GITHUB_TOKEN = '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'
